backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
openpyxl==3.1.2
langdetect==1.0.9
groq==0.4.1
anthropic==0.8.1
openai==1.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
